+++ src/sotkalib/func/__init__.py
@@ -1,3 +1,4 @@
+from . import importutil
 from .concur import asyncfn, asyncfn_or_raise
 from .cond import or_raise, suppress, type_or_raise
 from .defer import defer, defer_exc, defer_exc_mute, defer_ok
@@ -12,4 +13,6 @@ __all__ = [
 	"defer_exc_mute",
 	"defer_ok",
 	"defer",
+	# module
+	"importutil",
 ]