+++ src/sotkalib/serializer/impl/orjson.py
@@ -3,11 +3,9 @@ from typing import Any
 import orjson
 
 
-class ORJSONSerializer:
-	@staticmethod
-	def marshal(data: Any) -> bytes:
+class OrJsonSerializer:
+	def marshal(self, data: Any) -> bytes:
 		return orjson.dumps(data)
 
-	@staticmethod
-	def unmarshal(raw_data: bytes) -> Any:
+	def unmarshal(self, raw_data: bytes) -> Any:
 		return orjson.loads(raw_data)